+++ src/harmony/index.ts
@@ -1,0 +1,7 @@
+export { analogous } from './analogous';
+export { complementary } from './complementary';
+export { monochromatic } from './monochromatic';
+export { splitComplementary } from './split-complementary';
+export { tetradic } from './tetradic';
+export { triadic } from './triadic';
+